# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AcknowledgeAlertRequest:
    properties:
      comment:
        type: string
    type: object
...
  dto.AlertChannelResponse:
    properties:
      config:
        additionalProperties: true
        type: object
      createdAt:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  dto.AlertHistoryResponse:
    properties:
      acknowledgedAt:
        type: string
      acknowledgedBy:
        type: string
      channels:
        items:
          type: string
        type: array
      details:
        additionalProperties: true
        type: object
      errorMessage:
        type: string
      id:
        type: string
      message:
        type: string
      ruleId:
        type: string
      ruleName:
        type: string
      sentAt:
        type: string
      severity:
        type: string
      status:
        type: string
      triggeredAt:
        type: string
    type: object
  dto.AlertRuleResponse:
    properties:
      channels:
        items:
          type: string
        type: array
      conditions:
        items:
          $ref: '#/definitions/model.AlertCondition'
        type: array
      cooldown:
        type: integer
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      logic:
        type: string
      name:
        type: string
      severity:
        type: string
      template:
        type: string
      updatedAt:
        type: string
    type: object
  dto.AlertStatisticsResponse:
    properties:
      alertsBySeverity:
        additionalProperties:
          format: int64
          type: integer
        type: object
      alertsByStatus:
        additionalProperties:
          format: int64
          type: integer
        type: object
      recentAlerts:
        items:
          $ref: '#/definitions/dto.AlertHistoryResponse'
        type: array
      topAlertRules:
        items:
          $ref: '#/definitions/dto.TopAlertRule'
        type: array
      totalAlerts:
        type: integer
    type: object
...
  dto.CreateAlertChannelRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      enabled:
        type: boolean
      name:
        type: string
      type:
        enum:
        - webhook
        - slack
        - discord
        - dingtalk
        - wecom
        type: string
    required:
    - config
    - name
    - type
    type: object
  dto.CreateAlertRuleRequest:
    properties:
      channels:
        items:
          type: string
        minItems: 1
        type: array
      conditions:
        items:
          $ref: '#/definitions/model.AlertCondition'
        minItems: 1
        type: array
      cooldown:
        minimum: 1
        type: integer
      description:
        type: string
      enabled:
        type: boolean
      logic:
        enum:
        - AND
        - OR
        type: string
      name:
        type: string
      severity:
        enum:
        - low
        - medium
        - high
        - critical
        type: string
      template:
        type: string
    required:
    - channels
    - conditions
    - cooldown
    - logic
    - name
    - severity
    - template
    type: object
...
  dto.TestAlertChannelRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  dto.TopAlertRule:
    properties:
      count:
        type: integer
      ruleId:
        type: string
      ruleName:
        type: string
      severity:
        type: string
    type: object
...
  dto.UpdateAlertChannelRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      enabled:
        type: boolean
      name:
        type: string
    type: object
  dto.UpdateAlertRuleRequest:
    properties:
      channels:
        items:
          type: string
        type: array
      conditions:
        items:
          $ref: '#/definitions/model.AlertCondition'
        type: array
      cooldown:
        type: integer
      description:
        type: string
      enabled:
        type: boolean
      logic:
        enum:
        - AND
        - OR
        type: string
      name:
        type: string
      severity:
        enum:
        - low
        - medium
        - high
        - critical
        type: string
      template:
        type: string
    type: object
...
  model.AlertCondition:
    properties:
      duration:
        description: 持续时间（分钟）
        type: integer
      metric:
        description: qps, block_rate, error_rate, attack_count, etc.
        type: string
      operator:
        description: '>, <, >=, <=, ==, !='
        type: string
      threshold:
        description: 阈值
    type: object
...
  model.ErrResponseDontShowError:
    description: 错误的API响应标准格式,不展示 error
    properties:
      code:
        example: 400
        type: integer
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  model.SuccessResponse:
    description: 成功的API响应标准格式
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.SuccessResponseNoData:
    description: 成功的API响应标准格式,没有 data
    properties:
      code:
        example: 200
        type: integer
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
host: localhost:2333
info:
  contact:
    email: huahua1319873800@outlook.com
    name: API Support
    url: https://github.com/mingrenya/AI-Waf
  description: MRYa 应用防火墙管理系统 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/mingrenya/AI-Waf
  title: MRYa-WAF API
  version: "1.0"
paths:
...
  /alerts/channels:
    get:
      description: 获取所有告警渠道
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AlertChannelResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取告警渠道列表
      tags:
      - Alert
    post:
      consumes:
      - application/json
      description: 创建新的告警渠道
      parameters:
      - description: 创建渠道请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAlertChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertChannelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 创建告警渠道
      tags:
      - Alert
  /alerts/channels/{id}:
    delete:
      description: 删除指定的告警渠道
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 删除告警渠道
      tags:
      - Alert
    get:
      description: 根据 ID 获取告警渠道详情
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertChannelResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取告警渠道详情
      tags:
      - Alert
    put:
      consumes:
      - application/json
      description: 更新告警渠道信息
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新渠道请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAlertChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 更新告警渠道
      tags:
      - Alert
  /alerts/channels/{id}/test:
    post:
      consumes:
      - application/json
      description: 发送测试消息到指定渠道
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: string
      - description: 测试请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TestAlertChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 测试告警渠道
      tags:
      - Alert
  /alerts/history:
    get:
      description: 分页查询告警历史记录
      parameters:
      - description: 规则 ID
        in: query
        name: ruleId
        type: string
      - description: 严重级别
        in: query
        name: severity
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AlertHistoryResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取告警历史
      tags:
      - Alert
  /alerts/history/{id}:
    get:
      parameters:
      - description: 告警历史ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertHistoryResponse'
              type: object
      summary: 获取告警历史详情
      tags:
      - Alert
  /alerts/history/{id}/acknowledge:
    post:
      consumes:
      - application/json
      description: 确认已读告警
      parameters:
      - description: 告警历史 ID
        in: path
        name: id
        required: true
        type: string
      - description: 确认请求
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.AcknowledgeAlertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 确认告警
      tags:
      - Alert
  /alerts/rules:
    get:
      description: 获取所有告警规则
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AlertRuleResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取告警规则列表
      tags:
      - Alert
    post:
      consumes:
      - application/json
      description: 创建新的告警规则
      parameters:
      - description: 创建规则请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAlertRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertRuleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 创建告警规则
      tags:
      - Alert
  /alerts/rules/{id}:
    delete:
      description: 删除指定的告警规则
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 删除告警规则
      tags:
      - Alert
    get:
      description: 根据 ID 获取告警规则详情
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertRuleResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取告警规则详情
      tags:
      - Alert
    put:
      consumes:
      - application/json
      description: 更新告警规则信息
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新规则请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAlertRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 更新告警规则
      tags:
      - Alert
  /alerts/statistics:
    get:
      description: 获取告警统计信息
      parameters:
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertStatisticsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取告警统计
      tags:
      - Alert
...
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer {token} 格式进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"